feat: implement equaler; closes #43
